Stream incoming messages out to kafka as well as postgres
In order to communicate with the various indexes that will want to
consume a register, we'd like to use kafka.  This change causes the mint
to write all incoming messages to kafka as well as postgres.

The FunctionalTest is modified to spin up a JVM-local kafka (plus
zookeeper) instance and ensure that the message is written to both
postgres and kafka.

No thought has been put into error handling.  In particular, if the
postgres write succeeds but the kafka write fails, we could end up in an
inconsistent state between the data store and the kafka stream.