first stab. /latest.json presents last msg on topic
this is an MVP first attempt at a presentation app for consuming the
mint's kafka output.  There are two main classes:

 * PresentationResource, which defines a /latest.json endpoint
   - this can be extended with arbitrary representations in future by
     registering media types appropriately within Jersey
 * ConsumerRunnable, which consumes the kafka topic to determine the
   latest message

The ConsumerRunnable runs asynchronously in a separate thread, so it
communicates with the PresentationResource via an AtomicReference.

There is a FunctionalTest which creates a kafka topic, publishes a
message to it, and ensures the resource spits it out again after a
while.