reject messages which are not well-formed JSON
This updates the mint to reject messages which are not well-formed JSON.

Already the complexity of the message queue is rearing its head as we
need to decide what to do when a message is not well-formed: we want a
dead letter queue of some sort.  And how do we report back to the
original sender that there was a problem?

A better solution would be to remove RabbitMQ and make the entry
submission endpoint synchronous; the reason I haven't done this here is
it requires rewriting the bulk import script.