convert JSON to canonical form
We need to ensure that JSON is converted to a single canonical form, so
that any hashes we perform on it consistently give the same value for
the same semantic JSON content.

This commit ensures:

  - nonsignificant whitespace is trimmed
  - key order is consistent
  - unicode escape sequences are not used

This was loosely based on OLPC's consistent JSON format [1] except that
OLPC-Canonical-JSON requires newlines to be literal newlines, which is
actually banned by JSON as newlines are a control character (!).  So we
use escaped newlines instead.

If we want to be /really/ serious, we should probably use NFC
canonicalization on the Unicode strings too (!).  But this commit
doesn't attempt this.

[1]: http://wiki.laptop.org/go/Canonical_JSON