Present entries as real JSON from /latest.json
Before openregister/mint#7, we had no guarantee the incoming entries on
the kafka stream were legitimately JSON; so we were just treating them
as raw bytes to shove into a JSON string.  ie for message `foo`, we
would return JSON `["foo"]`.

This meant that when we got real JSON, we would embed it into a JSON
string.  That is, the message `{"foo":"bar"}` would return JSON
`["{\"foo\":\"bar\"}"]` rather than the desired `[{"foo":"bar"}]`.

This PR treats asserts that all incoming messages are JSON, and they
should be entered into a local postgres index.  Then, the
PresentationResource can get the latest message from the index.

This uses [dropwizard-jdbi][] for data access.  This provides an
advantage that the DBI objects it provides are managed by dropwizard and
will be started and shut down according to Dropwizard's lifecycle.  It
also allows database access through a nicely declarative interface.

[dropwizard-jdbi]: https://dropwizard.github.io/dropwizard/manual/jdbi.html