Show multiple entries in /latest.json
PR #1 started displaying JSON properly at /latest.json.  This PR ensures
that we can present more than one entry.

To do this, I needed to update FunctionalTest to clean the database each
time.  Because @BeforeClass methods run after TestRule setup, and we
needed to clean the database before starting the app, I needed to create
a new TestRule for cleaning the database.

Because the CleanDatabaseRule needs the db connection url, I've pulled
it out of the yaml file into the FunctionalTest class.