Remove dual writes problem from kafka sending
As the TODO in LocalDataStoreApplication said, there was a problem where
a write to Postgres could succeed but a write to Kafka could fail.  This
would result in an entry being missed out from the Kafka stream.
Changing the order of these would have the opposite effect -- we could
write entries to the stream which never were written to the database.

This changes the whole LogStream implementation.  Whereas before
LocalDataStoreApplication would write a message to PostgresDataStore and
LogStream, now it writes to PostgresDataStore and notifies LogStream
that there are new entries available.

LogStream then queries the database to find which new entries exist,
send all new entries, and record that they were sent, in a high water
mark table (called `streamed_entries`).  The table serves as a log of
which messages were sent at which time.

There are still failure modes where a message can be sent to kafka and
then an exception causes the new high water mark not to be recorded in
the streamed_entries table.  To deal with this, we now use the entry's
serial number as the Kafka primary key (which is *not* the register
primary key), and allow the consumer to deduplicate multiple messages
with the same primary key.  (We can also get Kafka to forget about old
messages with the same primary key, which is a safe operation since such
messages are guaranteed to be byte-for-byte identical).

As a side-effect, this makes the entry serial number visible to
consumers, which may become useful for constructing permalink URLs in
future.