Fixed performance of bulk load.
Previously the LoadHandler was inserting a single
entry at a time and for each of these was calling
kafka to process new entry.
This had a huge overhead.

The LoadHandler has been changed to batch update
entries to the database and call kafka once per
batch. This has reduced processing time of sample
data set (~34000 entries) from ~12m down to <10s.