Serialize SingleEntryView as JSON as well as HTML
We should use a single View instance for all of our representations.
This makes SingleEntryView a valid JSON representation so that Jackson
will serialize it, so we don't need separate methods to return
SingleEntryView for html and JsonNode for JSON.

This also introduces an Entry class to make serialization and
deserialization simpler.