deploy mint against RDS instead of docker postgres
In order to talk to RDS, we need to have configuration for where to find
the RDS endpoint and what username and password to use.  This commit
achieves this according to the strategy suggested on the amazon security
blog [1], where you store your configuration file in S3 and use IAM
Roles for EC2 instances to control access to the S3 bucket.

[1]: http://blogs.aws.amazon.com/security/post/Tx610S2MLVZWEA/Using-IAM-roles-to-distribute-non-AWS-credentials-to-your-EC2-instances