Name things consistently with our real-world discussions
We've had a number of sessions recently where we've talked about coming
up with concrete names for various concepts in registers.  These are
starting to crystallise into a common language.  This commit brings the
code into line with the names we are using.  In particular:

An /Entry/ is:

> an immutable mapping from /fields/ to /values/ created at a point in
> time, identified by a /serial number/

A /Register/ is:

> an append-only list of /entries/, with a defined set of /fields/, one
> of which is the /primary key/

A /Record/ is:

> the set of all /entries/ in a /register/ who share the same /primary
> key/ value

In this commit, I've also started using the word /content/ to refer to
the bag of data that an /entry/ contains, independently of its position
within the register.  (A hash identifies some content; two entries with
different serial numbers can have the same content and the same hash.)

I also have steered clear of using `entry` to refer to instances of
`Map.Entry`, using the term `keyValue` instead.  I don't particularly
like this -- maybe `mapEntry` would be better?  But I think it's
important that we don't overload our /entry/ term if we can avoid it.