Add ADR to document why we removed kafka
This commit adds an [Architecture Decision Record][] describing why we
chose to migrate away from kafka and zookeeper.  An Architecture
Decision Record is a document that records the context in which we made
a decision, the decision itself, and the forseen consequences of the
decision.  It is useful to record this information to ensure that if we
ever want to revisit a decision we can understand what the original
context was.

[Architecture Decision Record]: http://thinkrelevance.com/blog/2011/11/15/documenting-architecture-decisions