ensure csv renders all fields of a register
Previously, the CsvWriter would assume all entries had the same set of
fields and that it could get this by inspecting the first entry to
render.

This commit changes CsvWriter to get the list of fields to render from
the Register configuration class.

As a side bonus, a CsvWriter can now render an empty list.