Ensure Thymeleaf reads templates in UTF-8
There's a weird bug where thymeleaf seems to be getting confused by
utf-8 characters in the templates (eg ©, ’, –) and substituting them
with question marks (which is very naughty of it.)  The bug only
manifests remotely, not locally.  I wonder if this is an OS X/linux
difference.

Anyway this is an attempt to fix that.