hacky fix to deployment validation
the validation script fails because the app is looking for a "localhost"
register in the RegistersConfiguration (merged in b2dd5a99eaff0faee384f
in PR #78).  This hardcodes the validation script to pretend the
register is the country register just to get things working again.  The
longer-term solution is probably to use the dropwizard healthcheck
endpoint for validation.