better solution to deploy validation
This validate script works by hitting [dropwizard's healthcheck endpoint][hc].
The curl command is slightly simplified by using the `-f` flag to cause
curl to fail if the http status is 400 or greater.

Because the healthcheck endpoint is on a separate port that isn't
exposed to the world, we do some docker jiggery pokery to get the
container's IP address in order to hit port 8081.

[hc]: https://dropwizard.github.io/dropwizard/manual/core.html#health-checks