render register descriptive text as markdown
This should fix the country register at least!

This changes the tag from an `<h1>` as before to a `<div>` containing a
`<p>` tag.  The `<p>` tag is generated by the MarkdownProcessor so is
hard to get rid of.

This has changed the spacing of everything so could do with some design
help now.