create GIN index on jsonb column
Some of our queries are really slow because they are doing a full table
scan.  Because we're using jsonb, we can create a [gin index][] to make
these queries much more efficient.

There's an argument that the only things we care about in the data are
the primary key and the hash, so maybe we should instead pull those out
into separate columns and index that?  We don't need the fully
arbitrary querying capability of a general json index. :shrug:

[gin index]: http://www.postgresql.org/docs/9.4/static/datatype-json.html