Extract Indexer class and start in parallel
Because the DestinationDBUpdateDAO instance tries to create an index in
its constructor, and we create these objects in series in the main()
method, we can block the indexing of all registers while we try to
create an expensive index for a single big register.

This extracts that code into a new Indexer class, and starts these
indexers in parallel, so that we can start consuming one register while
creating a GIN index on another.

Once this is merged, we can revert #19.