Bug: where clause for must specify lower case index name to get the result.
When the application starts it creates the index if not exists using the check in code. the check was always returning false because of the query had index name ORDERED_ENTRY_INDEX_GIN all upper case in where clause while the index name created in db is 'ordered_entry_index_gin'.
 Changed all queries and table names in lower case to remove all confusion and always fetch db entites by there names in lower case.