Fix wonky table
The conditional here was put on the wrong tag.  This meant that if a
field was not a list (ie most of them), we'd put an empty `<ul></ul>`
tag on the page, which would mess up the layout.

The fix is to put the conditional on the `<ul>` tag, not on the `<th>`.