move HomePageView.getPublicBody() to ThymeLeafView.getCustodian()
We want to communicate who the custodian of the data is on many pages;
it makes sense that we should pull this up into ThymeLeafView, similar
to other globally-required things like register name.

While we're doing it, this also makes it clear that it's not just any
public body, but the custodian for this register, that this returns.