bump dropwizard to 0.9.1
There's a migration guide at [1] though it seems that we're not using
anything that needs migrating.

I added some exclusions to prevent things from the
`io.dropwizard.modules` group from pulling in `io.dropwizard`
dependencies, because they'll always be pulled in at 0.9.0, not 0.9.1.
As a result, we need to explicitly pull in dropwizard-jdbi at 0.9.1.

[1]: https://github.com/dropwizard/dropwizard/wiki/Upgrade-guide-0.8.x-to-0.9.x