support source maps for scss
This changes sassc to emit source maps, and checks in the generated
source map.

In order that the source maps make sense, we need the web server to
serve the original scss source files.  This is fixed by moving the
govuk_template, govuk_frontend_toolkit, and sass directories into
src/main/resources.

Once this is merged, we can view the original scss source files that
generated particular styles in web developer tools.