Logos for organisations
This uses code from alphagov/static to render logos alongside
organisation names on the front page and the top-right of the list and
single entry pages.

The ViewFactory is responsible for calling the GOV.UK organisations API
to fetch the logo, colour and wording for each organisation.

It seems that Thymeleaf strips newlines from text substituted on to the
page, so to get the newlines in the organisation branding to render
properly, I had to manually replace them with `<br/>` tags.