remove digital proof for a particular entry
This removes the "digital proof" section from the entry and record
pages (eg [here](http://school.openregister.org/school/402175) or
[here](http://school.openregister.org/entry/30568)).

We don't need this for a moment.  It's hardcoded and it's not something
we're intending to implement for our next milestone.  For the moment,
it's possible to validate integrity using the signed tree head on the
front page and validating the entire register against that tree head.

For larger registers, such as addresses, this method is not practical as
it means you can't prove a single entry comes from a register without
downloading and validating the entire register.  But for the moment,
it's okay.