Make register domain configurable for turtle prefix
We have a number of places throughout the codebase where we either
hardcode "openregister.org" as the register domain name, or we sniff it
from the Host header via RequestContext.

Since we're moving to our shiny new production register domain soon, we
should be able to make this configurable in one place.  This commit
creates the configuration item and uses it in one place only: the
@prefix line in the turtle output.  There are many other places that
could use this but they can come later :)