Make LinkValue's register domain configurable
This extends the previous commit to make LinkValue's register domain
configurable; previously it was hardcoded to openregister.org.
LinkValue is used in many places so it means that we are now using
configurable domains in many places, in particular including the turtle
representation's field values (as indicated in the changed turtle test
data).

There were three cases where it was not immediately possible to use a
configuration value for the LinkValue:

  - in HtmlViewSupport, which is a static method so can't access
    configuration
  - in PublicBody, which is a data class and so shouldn't have a
    reference to a configuration class
  - in HomePageView, where it should probably be passed in.

For the moment, these locations all hardcode the previous value of
".openregister.org" but they will need to be fixed.