make field headings register domain configurable
The field heading links on the /records, /entries, and individual record
and entry pages were hardcoded to link to
http://field.openregister.org.  This change uses the value from the
configuration class instead.

This is done by putting the registerDomain in to the RequestContext
class, but this is a bit ugly because the register domain has nothing to
do with the request context.  I only did this for convenience;
RequestContext is a place I can inject things rather than having to pass
the configuration all the way through from the ViewFactory through to
the ThymeLeafView superclass.  There's probably a better way to do this.