Create functional test for register resource.
The test makes sure that the register.json resource contains the entry of the register in the Register register. To ensure that the Register register (postgres) and the register details in the address register (registers.yaml) are consistent the registers.yaml config file was updated to ensure it contains serial-number (as this serial-number comes from the database for the Register register). Also changed the DbSupport class to allow serial-number to be specified by the tests when putting test data in the db, rather than using an arbitrary counter to create the serial-number. This makes sure both Register register and Address register are using consistent data for registers.