Change RegisterResourceFunctionalTests to test that last-updated for the register is extracted from the leaf_input correctly.
DBSupport was changed such that publishMessages() will return the timestamp for the entry with the greatest serial_number that is inserted into the ordered_entry_index table before it is converted into a CT leaf_input. The test then makes sure that this timestamp is returned correctly formatted in /register.json.