Throw 500 error if CT server doesn't like our request
Previously, if the CT server returned a 400 Bad Request, we would just
return that to the caller.  However on reflection if we are making a
request to the CT server which is malformed in some way, that indicates
that we have a bug in mint that needs to be fixed.  The correct response
in this case is 500 Internal Server Error, to demonstrate that the
problem lies within mint rather than with whoever is submitting
something to mint.