support different phases
Currently we hardcode our phase banner to "ALPHA".  We should support
displaying a different phase depending on the phase of the register from
the register configuration.  A phase of "alpha" gives the alpha banner,
a phase of "beta" gives the beta banner, and a phase of anything
else (such as "live") removes the banner completely.