bump registers to beta
This is done in a hacky way to get it out the door.  PR #169 introduced
behaviour to allow the UI to reflect the phase of the registers; while
openregister/registry-data#7 updated the register/field/datatype
registers with the actual phases.

This bumps the phase values for registers.yaml to match those in
openregister/registry-data#7.

It ignores fields.yaml and datatypes.yaml because they don't affect the
UI at this stage.

At some point we should start pulling in this data from environmental
configuration and not from the jar file.