Fix connection pooling in tests
I had problems where RepresentationsTest was throwing exceptions saying
that remaining connections were reserved for superuser.  It seemed that
we were running out of connections.

The old DBSupport.createConnection() method created a pooled DataSource,
which was good, but it created a new separate DataSource for each
connection, which kind of negated the point of a connection pool.

This commit changes DBSupport to have a single static pooled DataSource,
and give connections out from that single pool.

For the main() method which needs to talk to a different URL (called
from the gradle loadSchoolData task), it can't now reuse the
createConnection() method because it needs a DataSource pointing at a
different URL.  So it needs to create its own DataSource.  It still only
creates one DataSource so that's okay.