Using total_entries table last_updated column to render last_updated time of a register.
We are not going to use x-json sever (refer: https://github.com/openregister/mint/pull/64) which allows us to remove leaf_input column completely from ordered_entry_index table. last_update column in total entries contains the time when the last entry was added in to the register which can be used now for this purpose.