try more up-to-date python from ppa
The main `precise` repositories only go up to python 3.2.  However `pip`
now does not support 3.2 and requires >= 3.3.  We need `pip` in order to
install and run the conformance tests.

We can instead install a later python from a ppa.  Travis has a
whitelist of acceptable ppas to install software from.

The "deadsnakes" alias [1] refers to the whitelisted
ppa:fkrull/deadsnakes [2] which provides lots of python versions for
lots of ubuntu versions.  We use this to pull in python 3.5 instead.

[1]: https://github.com/travis-ci/apt-source-whitelist/blob/dce2f6b8fd2ccaf4fcd41dbe2f17c69e80d6d6a9/ubuntu.json#L72-L76
[2]: https://launchpad.net/~fkrull/+archive/ubuntu/deadsnakes