need to also install pyvenv separately
although `venv` is part of the standard python 3 distribution (since
3.3), the deadsnakes packages split it into a separate package.

Also, they don't provide a `pyvenv` alias for the binary so it can only
be referred to as pyvenv-3.5.  I've changed the gradle file to use this
name.  I don't like this solution though, it feels brittle.  Longer term
I feel it'd be better to configure travis to create a `pyvenv` alias
somehow.