Don't try to createdb in create_schema.sh
I tested this functionality by dropping my `ft_presentation` db and
trying to run the build.  The app failed to start, because the
ft_presentation db didn't exist.

The problem is that `create_schema.sh` is run after the app is started,
so creating the db at this point is too late to allow the app to run.

We already have the `./go` script to create this db, which is referenced
from the README, so I think we can just remove the createdb call from
create_schema.sh.