Use registers & fields config from S3
Similar to openregister/mint@717bcdfa4f255229520.  The files are already
available and in use by the mint in each environment.  We already
introduced the ability to configure a path for the registers yaml file
in e00c2402e20fb0.  This just updates the deployment scripts to be able
to actually get the files from S3 and point presentation at them.