Render copyright text as markdown
The copyright field is defined to be markdown, and for the postcode
register the data is markdown.  This updates the copyright text to be
rendered as markdown.

Copyright text can appear in the main body or on the footer.  I've
written tests to catch both cases.

We are already rendering markdown on the front page from the register
`text` field, so this was relatively simple to add.