Add Strict-Transport-Security header
We should serve HSTS headers since register.gov.uk is only available
over HTTPS.  We already serve HSTS on the apex domain and have been
added to the preload list; this adds the header on each individual
foo.register.gov.uk domain too, which means we will support HSTS for
browsers which don't use the preload list.

This is slightly questionable in non-production environments because
according to [RFC 6797 §7.2][], a server MUST NOT send
Strict-Transport-Security headers over a non-secure transport, and
non-prod environments are available over plain http.  However, the RFC
also says that a user agent MUST ignore an STS header over an insecure
transport, so this isn't such a big deal.  Given that this only affects
non-production environments I'm not too concerned about this.

I haven't written a test for this; longer-term, it's not clear that the
application is necessarily the right place to add this header.  (I'm
open to argument that we should test it anyway though).

[RFC 6797 §7.2]: https://tools.ietf.org/html/rfc6797#section-7.2