Convert item table content to JSONB
The mint just stored item content as `bytea`.  That was okay for mint,
but a merged app will need to do JSON queries against item contents (eg
to find items by primary key), so we should use an item table based on
JSONB instead.

This means we no longer store the original canonical representation of
the item in the database but it is (or should be) deterministically
produced by the canonical json mapper.