remove code to update presentation's entry/item tables
The indexer has code to:

 1) copy mint's entry and item tables over to presentation
 2) create the `current_keys` table, which is used for doing efficient
    record lookups

Once the apps are merged, task 1) is no longer necessary.

This commit removes the code from DestinationDBUpdateDAO which was to do
with 1).  It also extracts any further dependency on EntryUpdateDAO into
a new subclass, with the intention that the new subclass can eventually
be removed.