convert verifiablelog test to a unit test
The old VerifiableLogResourceFunctionalTest was tricky to convert to a
new-style functional test that mints entries through the Mint API
because that meant that it didn't have control over the exact time
that the entries get minted, and so cannot predict the correct root
hash to expect.

By converting it to a unit test, we can control all the things we need to
while still testing the same behaviour. As a bonus, the test runs faster.

Hurrah!

Paired with @chr1sdi