change EntryMapperTest to go via test postgres db
The old test uses a mocked ResultSet. This made me uncomfortable:

 - "Only mock types you own" - §4.1 of "Mock Roles, not Objects"
   http://www.jmock.org/oopsla2004.pdf
 - the mocked getTimestamp() methods didn't accurately reflect the
   contract of ResultSet: in particular they completely ignored the
   Calendar argument; meaning that we never tested that a correct
   Calendar instance with the right timezone was passed.

This new test integrates against an actual postgres database directly in
the test.  The .travis.yml file was already creating the
`test_openregister_java` database but nothing was using it, so i decided
to steal it for this test.

We insert an entry and expect that the resulting EntryQueryDAO correctly
interprets the timestamp column value.  This implicitly calls the
EntryMapper class.