build fat jar with shadow plugin
We have a homegrown method for building a fat jar.  This works as far as
copying files into the same place and bundling them together, but when
building fat jars there are other tasks that need to be done, such as
merging service metadata information files together.

This bit me when I tried to configure logging for heroku, because when I
tried to specify a logger with `type: console` I got a weird error,
which I eventually worked out was because the service metadata hadn't
been handled properly.

By moving to the shadowJar plugin, we can perform this task simply by
using the mergeServiceFiles() method.

Longer term, we should move to only one method for producing a fat jar.