remove dynamic register proof from home page
For large registers, each time we redeploy we'll trash the in-memory
cache of merkle tree nodes.  If we have the root hash on the homepage,
then each time we redeploy, we won't be able to render the homepage
until the in-memory cache has been rebuilt.

Reverting back to the static mockup for the time being.