Introduce Register interface
Introduce a Register interface, to be an abstraction layer representing the
high-level operations you can do with a register.

There is a PostgresRegister implementation, backed by the current
postgres database stuff.

This is not complete, nor is it efficient. We've lost support for
inserting items, entries and records in batch, resulting in a much less
efficient (slower) sql update because we're bounded by network round-trip-time.