add mintItems() method
I wonder if really the operation you can do is to transactionally add
multiple items to a register.  Therefore, we should add a plural
mintItems() method which allows adding multiple items, but which means
that the Register object (not the caller) manages the transaction.

I wonder if the same pattern could extend to a method like
`slurpExternalRepresentation(Iterable<Command>)` for slurping in (and
validating) data from an external representation.