extract schema creation to separate SchemaCreator class
The more I think about it, the more I feel that the DAOs aren't the
right place to manage db schemas.  Longer term, we should have a proper
migration tool (such as dropwizard-migrations) but at least for the
moment we can extract the schema creation to a single separate class,
distinct from the query/update DAO classes.