Convert RegisterDAO to onDemand SQL object
Previously I was copying the prior approach we had to EntryStore - which
was to open a Handle in the RegisterApplication class and assign it to
the SQL Object.  This has the problem that it only leaves one Handle for
the whole application; I presume this therefore means the app can only
use a single database connection, which means if someone is loading a
large batch then nobody else can make any requests concurrently.

This commit converts RegisterDAO to a class suitable for using via the
.onDemand() method; in particular, it expects to get given a fresh
Handle for each request, and will pass this Handle on to any subsequent
DAO instances that it needs to use.

This will result in more object churn as we'll be creating and
destroying multiple DAO objects on each request.  I suspect this is fine
but we can address this as a separate problem later.