run proof queries in transactions
under the hood, the proof queries call
EntryQueryDAO.entriesIteratorFrom() to open a cursor, possibly multiple
times.  We should probably ensure we're doing these queries at a
snapshot in time.

We're not currently clear on the difference between a connection with
autocommit == false, and a connection with an explicit transaction.  But
at least we know a transaction will do the right thing.