specify db url in test-app-config.yaml
Apparently I introduced this PLACEHOLDER value in 5deab9df2b60262df7 as
a way to DRY up some test code.  But this resulted in some problems:

  - every single time we created a DropwizardAppRule, we had to specify
    `database.url`, and we always set it to the same value
  - the test-app-config.yaml file couldn't be used on its own without
    any ConfigOverrides.  In particular, the idea-run-config.gradle had
    to use the RegisterApplicationRunner class in order to run the app
    with the test-app-config.yaml and the right database url.

This commit re-hardcodes the URL in the yaml file again.  Then it
deletes a whole bunch of wiring around `database.url`, and changes the
idea-run-config.gradle so that RegisterRunner and
FunctionalTestRegisterRunner are the same task except for the
configuration file.