Introduce RegisterRule abstraction
Lots of our tests want to create a register app server with a given
register name configuration, fire some requests against it, possibly
wipe it before tests, and then tear down at the end.

This introduces a RegisterRule class which wraps DropwizardAppRule and
provides an interface for loading data in and making requests against
the register.  It applies it to RepresentationsFunctionalTest to remove
a bunch of accidental complexity from that class.