Remove some usages of DBSupport
This migrates code to use `RegisterRule.mintLines()` rather than
`DBSupport.publishEntries()` wherever possible.

I removed
RegisterResourceFunctionalTest.registerJsonShouldContainEntryViewAddressRegister
because a) it was hard to keep it as it requires running the register in
two different modes in the same test class, and b) I'm not sure it's
actually doing anything useful.