Remove DBSupport and all who use it
The `DBSupport` class was a way of stuffing data into the database
rather than going through the register interface properly.  It was a
useful way of testing the `presentation` app back when we had three
separate apps which integrated on the database.

However, now we have one app which wholly owns the database, so it's a
bit rude for the tests to insert things into the database themselves
when they can use the `/load` or `/load-rsf` endpoints to load data in.

The last places where we were still doing this were in gradle tasks
which called the `DBSupport` class as a main class.  These can now be
replaced with calls to `curl` to load in data via the `/load-rsf`
endpoint.  Having done this, `DBSupport` and a lot of its accompanying
code can be deleted. :tada: